+++ src/shared/api/dto/index.ts
@@ -1,3 +1,3 @@
 export * from './hero-config.dto';
-export * from './market-listing.dto';
+export * from './market-buy.dto';
 export * from './skin-config.dto';