+++ src/worker.module.ts
@@ -2,12 +2,14 @@ import { EkConfigService, SdkModule } from '@earnkeeper/ekp-sdk-nestjs';
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { MarketBuyModule } from './feature/market-buy/market-buy.module';
+import { MarketRentModule } from './feature/market-rent/market-rent.module';
 import { QueueModule } from './shared/queue/queue.module';
 
 export const MODULE_DEF = {
   imports: [
     MongooseModule.forRootAsync({ useClass: EkConfigService }),
     MarketBuyModule,
+    MarketRentModule,
     QueueModule,
     SdkModule,
   ],