+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [0.1.1] - 2022-04-19
+
+### Added
+
+- ✨ Profit calculations to Buy Heroes page
+- ✨ New Page: Rent Heroes, the thetan arena hero rentals marketplace, with profit calculator!
+  
 ## [0.1.0] - 2022-04-18
 
 ### Added